chore: split preview-related source to distinct folder
