+++ src/preview/get-component.ts
