+++ src/preview/highlight-area.ts
