+++ src/preview/pattern-id-to-webpack-name.ts
